feat: list transactions and create transactions

+++ server/kadi/src/main/java/com/codewithandrey/kadi/wallet/dto/CreateWalletRequest.java
@@ -3,6 +3,6 @@ package com.codewithandrey.kadi.wallet.dto;
 public record CreateWalletRequest(
         String name,
         com.codewithandrey.kadi.wallet.Currency currency,
-        Integer balance
+        Long balance
 ) {
 }
